fix time cost format

+++ src/user.py
@@ -28,7 +28,7 @@ def app_open(files):
                     value['ip'] = line_list[0]
                     value['country'] = ip_location(value['ip'])
                     value['status'] = line_list[7]
-                    value['time_cost_new'] = int(line_list[-3])
+                    value['time_cost_new'] = float(line_list[-3])
                     value['Android'] = 'Android' in line
                     value['iPhone'] = 'Club_Factory_UIWebView' in line
                     if 'Android' in line:
